pyproject.toml
[tool.poetry]
name = "asyncdbdriverseveralsources"
version = "0.1.0"
description = ""
authors = ["rpopov <rpopov@web-bee.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
sqlalchemy = "^2.0.15"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pytest-postgresql = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
